refactor hooks and use hook for typechecking
